Omit `main_app` prefix for url helpers
Currently, when overriding the views for the Pointless Feedback engine,
the main app application layout (and any included partials) needs
modified to prefix `main_app` in front of url helpers (i.e. `foo_path`).
This commit adds a method_missing to pass methods that appear to be url
helpers and that the `main_app` defines to the `main_app`.